add base method to parser
